Update for new kernel ABI
There are a couple of things in here that don't have an exact equivalent
in the new ABI, but these changes are enough to make a bunch of tests
work.

+++ src/libwrap/nvhost.c
@@ -168,7 +168,7 @@ static void nvmap_file_enter_ioctl_write(struct nvmap_file *nvmap,
 
 	dest = (uint8_t *)handle->buffer + op->offset;
 	end = (uint8_t *)handle->buffer + handle->size;
-	src = (uint8_t *)op->addr;
+	src = (uint8_t *)(uintptr_t)op->addr;
 
 	for (i = 0; i < op->count; i++) {
 		print_hexdump(stdout, DUMP_PREFIX_NONE, "  ", src,
